+++ src/main/java/duchess/tasks/Event.java
@@ -1,3 +1,5 @@
+package duchess.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
