+++ src/main/java/duchess/tasks/Task.java
@@ -1,3 +1,5 @@
+package duchess.tasks;
+
 public class Task {
     protected String description;
     protected boolean isDone;
