+++ src/main/java/duchess/tasks/ToDo.java
@@ -1,4 +1,6 @@
-public class ToDo extends Task{
+package duchess.tasks;
+
+public class ToDo extends Task {
 
     public ToDo(String description) {
         super(description);
