+++ src/main/java/duchess/util/DuchessException.java
@@ -1,3 +1,5 @@
+package duchess.util;
+
 public class DuchessException extends Exception {
     public DuchessException() {
