+++ src/main/java/duchess/util/Parser.java
@@ -1,3 +1,5 @@
+package duchess.util;
+
 public class Parser {
     public Parser() {
