+++ src/main/java/duchess/util/TaskList.java
@@ -1,7 +1,12 @@
+package duchess.util;
+
+import duchess.tasks.Task;
+
 import java.util.ArrayList;
 public class TaskList {
     private ArrayList<Task> taskList;
     public TaskList() {
+        this.taskList = new ArrayList<>();
     }
     public TaskList(ArrayList<Task> taskList) {
         this.taskList = taskList;
